Determine year in SQLite URL automatically

+++ sqlite/build.sh
@@ -17,7 +17,19 @@ version_dir=$recipe_dir/build/$version
 mkdir -p $version_dir
 cd $version_dir
 src_filename=sqlite-autoconf-$version_int.tar.gz
-wget -c https://www.sqlite.org/$year/$src_filename
+
+current_year=$(date +%Y)
+min_year=2022
+for year in $(seq $current_year -1 $min_year); do
+    if wget -c https://www.sqlite.org/$year/$src_filename; then
+        break
+    fi
+done
+
+if [ ! -f $src_filename ]; then
+    echo "Couldn't find URL for $src_filename" >&2
+    exit 1
+fi
 
 build_dir=$version_dir/$HOST
 rm -rf $build_dir
